Fix ignrance of credential files for model registry
